feat: CLIN-3267 First draft
Temporary commit to avoid losing work locally.

This code works locally with minikube for test_nextflow_operator and etl_variant_annotation.
It is not tested on etl.py.

It reuse a configuration mechanism similar as in cqdg
Key modifications in comparison to cqdg:
- service_account_name specified in BaseConfig instead KubeConfig
- shallow copy in build_operator and partial methods in BaseConfig
- BaseConfig.args renamed to BaseConfig.append_args

There are a few todos in the code to address.